Fix game launching client side

+++ client/Splendor Client/Assets/Scripts/Game/MainGame/Player.cs
@@ -6,7 +6,7 @@ using UnityEngine.UI;
 public class Player : MonoBehaviour
 {
     private string username;
-    private int points = 0;
+    private long points = 0;
     private List<Card> acquiredCards = new List<Card>(), reservedCards = new List<Card>();
     private List<Noble> acquiredNobles = new List<Noble>(), reservedNobles = new List<Noble>();
     private CardGemValue tokensAcquired = new CardGemValue();
@@ -28,7 +28,7 @@ public class Player : MonoBehaviour
         return username;
     }
 
-    public void SetPoints(int points) {
+    public void SetPoints(long points) {
         this.points = points;
         //update display
         MultiplayerInfoPanel infoPanel = this.GetComponent<MultiplayerInfoPanel>();
@@ -42,7 +42,7 @@ public class Player : MonoBehaviour
         }
     }
 
-    public int GetPoints()
+    public long GetPoints()
     {
         return points;
     }
